fix segfault with empty input

+++ modules/cap3.nf
@@ -5,11 +5,14 @@ process cap3{
     tuple val(assembly_name), path(assembly)
 
     output:
-    tuple val("${assembly_name}-cap3"), path("${assembly_name}-cap3.fasta")
+    tuple val("${assembly_name}-cap3"), path("${assembly_name}-cap3.fasta"), optional true
 
     script:
     """
-    cap3 ${assembly} > cap3.log 2> cap3.log
-    cat ${assembly}.cap.contigs ${assembly}.cap.singlets > ${assembly_name}-cap3.fasta
+    if [ -s ${assembly} ]
+    then
+        cap3 ${assembly} > cap3.log 2> cap3.log
+        cat ${assembly}.cap.contigs ${assembly}.cap.singlets > ${assembly_name}-cap3.fasta
+    fi
     """
 }
